+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.0.0-beta.2 - 2019-01-10
+
+### Added
+- Added console command to trigger resaving of caches.
+- Added web controller to trigger resaving of caches.
+
 ## 1.0.0-beta.1 - 2018-12-10
 
 ### Added